+++ scripts/lichessTools-init.js
@@ -111,6 +111,7 @@ lichessTools.loadTool(LiChessTools.Tools.ProfileTimeline);
 lichessTools.loadTool(LiChessTools.Tools.PlayWheel);
 lichessTools.loadTool(LiChessTools.Tools.DailyQuote);
 lichessTools.loadTool(LiChessTools.Tools.PuzzleDownload);
+lichessTools.loadTool(LiChessTools.Tools.GameListOptions);
 //lichessTools.loadTool(LiChessTools.Tools.DownloadGames);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);