+++ HISTORY.md
@@ -2,10 +2,13 @@ History of features added to LiChess Tools in time (reversed order)
 
 18 Oct
 
-- puzzle download
+- board coords on exported board image
+- game list options
 
 17 Oct
 
+- puzzle download
+- suggest chapter name from mainline moves
 - daily quote
 
 15 Oct